refactor(worktree): consolidate worktree-path slug rule in worktree-core
Two implementations of the worktree-path slug rule diverged:

- `orchard-worktree::cmd::worktree_path_for` →
  `<repo_root>/.worktrees/<branch.replace('/', '-')>`
- `orchard::paths::derive_local_worktree_path` →
  `<parent(repo_root)>/worktrees/worktree-<sanitized-slug>`

`orchard new feat/foo` produced one path; the TUI dialog produced a
different one for the same branch, so worktrees created via the CLI
weren't where the TUI expected them.

Hoist `worktree_path_for(repo_root, branch) -> PathBuf` into
`worktree-core::path` (the existing source of truth for worktree
mutations) and converge:

- `orchard-worktree::cmd::new` calls `worktree_core::worktree_path_for`.
- `orchard::tui::start_create_worktree` calls
  `worktree_core::worktree_path_for` instead of
  `paths::derive_local_worktree_path`.
- Drop `derive_local_worktree_path` (and the local `worktree_path_for`
  in `orchard-worktree`) — the worktree-core helper is the single
  source of truth.

`paths::sanitize_branch_slug` stays — it's used by remote-registry-entry
slug naming, which is a separate concern from the worktree path.

Behavioural change: TUI dialog now creates worktrees at
`<repo_root>/.worktrees/<branch-with-slashes>`. The de-facto layout
already uses this convention; the TUI's old convention produced no
files in this repo.

Closes #446
