+++ crates/orchard-worktree/src/cmd/mod.rs
@@ -10,19 +10,7 @@ pub mod path;
 pub mod prune;
 pub mod rm;
 
-use std::path::PathBuf;
-
 use anyhow::{Result, anyhow};
-
-/// Resolve the path that a new worktree for `branch` should live at.
-///
-/// Convention: `<repo_root>/.worktrees/<branch-with-slashes-replaced>`.
-/// Slashes become hyphens because `.worktrees/foo/bar` would imply a
-/// directory hierarchy that doesn't exist.
-pub fn worktree_path_for(repo_root: &str, branch: &str) -> PathBuf {
-    let slug = branch.replace('/', "-");
-    PathBuf::from(repo_root).join(".worktrees").join(slug)
-}
 
 /// Resolve a worktree by branch name to its absolute path.
 ///
@@ -34,33 +22,4 @@ pub fn resolve_worktree_path(branch: &str) -> Result<String> {
         .find(|t| t.branch.as_deref() == Some(branch))
         .map(|t| t.path.clone())
         .ok_or_else(|| anyhow!("no worktree found for branch '{branch}'"))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn worktree_path_for_simple_branch() {
-        assert_eq!(
-            worktree_path_for("/repo", "feature").to_string_lossy(),
-            "/repo/.worktrees/feature"
-        );
-    }
-
-    #[test]
-    fn worktree_path_for_slashed_branch_replaces_with_hyphen() {
-        assert_eq!(
-            worktree_path_for("/repo", "feature/foo").to_string_lossy(),
-            "/repo/.worktrees/feature-foo"
-        );
-    }
-
-    #[test]
-    fn worktree_path_for_double_slash_branch() {
-        assert_eq!(
-            worktree_path_for("/repo", "issue123/feature/sub").to_string_lossy(),
-            "/repo/.worktrees/issue123-feature-sub"
-        );
-    }
 }
