+++ crates/worktree-core/src/lib.rs
@@ -31,11 +31,13 @@
 pub mod create;
 pub mod destroy;
 pub mod list;
+pub mod path;
 pub mod prune;
 pub mod repo;
 
 pub use create::{CreateOutcome, create_worktree};
 pub use destroy::remove_worktree;
 pub use list::{WorktreeEntry, list_worktrees, parse_porcelain, worktree_has_conflicts};
+pub use path::worktree_path_for;
 pub use prune::prune;
 pub use repo::{find_repo_root, get_repo_name};
